Split test case
